pyproject.toml
[project]
name = "epicc"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.55.0",
    "pandas>=2.3.0",
    "numpy>=2.4.0",
    "openpyxl>=3.1.0",
    "pydantic>=2.12.5",
    "ruamel-yaml>=0.19.1",
    "plotly>=6.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.6",
    "pytest>=8.0.0",
    "mypy>=1.20.0",
    "types-pyyaml>=6.0.12.20250915",
    "pandas-stubs>=3.0.0.260204",
    "types-openpyxl>=3.1.0.20240106",
    "plotly-stubs>=0.1.3",
    "playwright>=1.58.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.mypy]
mypy_path = "src"
explicit_package_bases = true

[tool.stlite]
title = "EPICC Cost Calculator"
mount_dirs = ["src"]
text_suffixes = [".py", ".yaml", ".yml", ".css", ".js", ".html"]
css_url = "https://cdn.jsdelivr.net/npm/@stlite/browser@1.4.0/build/stlite.css"
js_url = "https://cdn.jsdelivr.net/npm/@stlite/browser@1.4.0/build/stlite.js"
